+++ changelog.md
@@ -1,3 +1,6 @@
+## Fabric
+- fix bug when killing an animal, the exp was dropped every time, even with magnetic enabled
+
 ## Paper
 
 - significantly improves break-chained block handling